refactor(shields): isolate mutable config repair
(cherry picked from commit 717d6ef70feb4df1b807a6597c9635eae5709368)
